% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_param.R
\name{autotune_param}
\alias{autotune_param}
\title{The function to control the acceptance rate of the Metropolis steps by tuning tau.}
\usage{
autotune_param(
  draw_param_list,
  tune_param,
  accpt_list,
  tune_nterm,
  tune.incr,
  tune.accpt.prop.lower,
  tune.accpt.prop.upper,
  ...
)
}
\arguments{
\item{draw_param_list}{The list of all iterates of A or B till the current iteration.}

\item{tune_param}{The matrix of tau, i.e. the standard deviations of the proposal distribution of A or B.}

\item{accpt_list}{The list of the acceptance for all iterates of A or B till the current iteration.}

\item{tune_nterm}{The iteration after which we start to tune tau.}

\item{tune.incr}{The adjustment proportion for tuning tau.}

\item{tune.accpt.prop.lower}{The ideal lower bound for the acceptance rate of the Metropolis steps.}

\item{tune.accpt.prop.upper}{The ideal upper bound for the acceptance rate of the Metropolis steps.}

\item{...}{Other parameters that may be useful.}
}
\description{
The function to control the acceptance rate of the Metropolis steps by tuning tau.
}
