% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{BIC}
\alias{BIC}
\title{BIC calculation for the output of \code{SIMP()}}
\usage{
BIC(SIMP.fit, r, pc, pd, p2, dx, dy, samp.size, combined = T, Chain.no = 1)
}
\arguments{
\item{SIMP.fit}{The output of \code{SIMP()}.}

\item{r}{Dimension of response Y.}

\item{pc}{Dimension of X1C, the continuous part of the predictors of interest.}

\item{pd}{Dimension of X1D, the discrete part of the predictors of interest.}

\item{p2}{Dimension of X2, predictors of not main interest.}

\item{dx}{Partial predictor envelope dimension of SIMP.}

\item{dy}{Partial response envelope dimension of SIMP.}

\item{samp.size}{Sample size of the datasets.}

\item{combined}{Logical. Indicate whether posterior samples (after burn-in) from all chains 
should be combined to calculate BIC. If there is only one chain, then there is no difference 
by indicating this parameter to be TRUE or FALSE.}

\item{Chain.no}{If combined = FALSE, we need to indicate which chain should be used for the 
calculation of BIC. Chain.no must be an integer between 1 and the number of chains.}
}
\description{
BIC calculation for the output of \code{SIMP()}
}
