% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{build_tree}
\alias{build_tree}
\title{Tree building function for NUTS}
\usage{
build_tree(
  A,
  j,
  aa,
  u,
  v,
  jj,
  eps,
  A00,
  aa0,
  lpd_func,
  grad_lpd_func,
  M_diag,
  Delta_max = 1000,
  ...
)
}
\arguments{
\item{A}{Parameter A or B.}

\item{j}{The index for which column of A or B to be updated.}

\item{aa}{The momentum vector of the j-th column of A or B.}

\item{u}{a random number used in NUTS.}

\item{v}{a random number from discrete Unif{-1, +1} used in NUTS.}

\item{jj}{The index for the layer of the tree.}

\item{eps}{Stepsize.}

\item{A00}{A or B from the previous iteration.}

\item{aa0}{The momentum vector from the previous iteration.}

\item{lpd_func}{The log full conditional posterior density of A or B.}

\item{grad_lpd_func}{The function to calculate the gradients of 
the log full conditional posterior density of A or B.}

\item{M_diag}{Covariance matrix for our matrix parameter A or B.}

\item{Delta_max}{The Delta_max parameter used in the Algorithm 6 of NUTS.}

\item{...}{Other parameters that may be useful.}
}
\description{
Tree building function for NUTS
}
