% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bura_Cook.R
\name{Bura_Cook}
\alias{Bura_Cook}
\title{Bura_Cook estimator for the estimation of rank(beta_{1C})}
\usage{
Bura_Cook(X1C, X1D, X2, Y, sig.level = 0.05)
}
\arguments{
\item{X1C}{Design matrix of the continuous part of the predictors of interest.
Must have the same number of rows as \code{Y}.}

\item{X1D}{Design matrix of the discrete part of the predictors of interest.
Must have the same number of rows as \code{Y}.}

\item{X2}{Design matrix of the nuisance predictors. Must have the same number of rows as \code{Y}.}

\item{Y}{Response matrix. Must have the same number of rows as \code{X}.}

\item{sig.level}{Significance level of the sequence of Chi-squared tests called in 
Bura-Cook estimator.}
}
\description{
Bura_Cook estimator for the estimation of rank(beta_{1C})
}
\examples{
\dontrun{
library(SIMP)
library(Renvlp)  # Load Renvlp package only for wheatprotein dataset.
data(wheatprotein)
set.seed(1)
d <- Bura_Cook(X1C = wheatprotein[, 4:5], X1D = matrix(wheatprotein[, 8], ncol = 1), 
                X2 = wheatprotein[, 6:7], Y = wheatprotein[, 1:3], sig.level = 0.05)
}
}
