% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{check_NUTS}
\alias{check_NUTS}
\title{Conditions checking function used in NUTS.}
\usage{
check_NUTS(s, A_plus, A_minus, aa_plus, aa_minus, j)
}
\arguments{
\item{s}{An output from build_tree()}

\item{A_plus}{An output from build_tree()}

\item{A_minus}{An output from build_tree()}

\item{aa_plus}{An output from build_tree()}

\item{aa_minus}{An output from build_tree()}

\item{j}{The index for which column of A or B to be updated.}
}
\description{
Conditions checking function used in NUTS.
}
