% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{CV}
\alias{CV}
\title{Bayesian CV}
\usage{
CV(SIMP.fit.full, Y, X1C, X1D, X2, dx, dy, K_CV, n.iter, correction = FALSE)
}
\arguments{
\item{SIMP.fit.full}{The output of \code{SIMP()} fitted on the full dataset. This input is used as
correction, and the posterior samples from all chains will be combined.}

\item{Y}{Response matrix. Must have the same number of rows as \code{X}.}

\item{X1C}{Design matrix of the continuous part of the predictors of interest.
Must have the same number of rows as \code{Y}.}

\item{X1D}{Design matrix of the discrete part of the predictors of interest.
Must have the same number of rows as \code{Y}.}

\item{X2}{Design matrix of the nuisance predictors. Must have the same number of rows as \code{Y}.}

\item{dx}{Partial predictor envelope dimension. Must be an integer between 0 and \code{ncol(X1C)}.}

\item{dy}{Partial response envelope dimension. Must be an integer between 0 and \code{ncol(Y)}.}

\item{K_CV}{Fold for CV}

\item{n.iter}{Number of Markov chain iterations to run in each chains, for the fitting in CV.
*Includes burn-in*.}

\item{correction}{Whether correction should be applied.}
}
\description{
Bayesian CV
}
