% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{DIC}
\alias{DIC}
\title{DIC calculation for the output of \code{SIMP()}}
\usage{
DIC(SIMP.fit, Y, X1C, X1D, X2, dx, dy, combined = T, Chain.no = 1)
}
\arguments{
\item{SIMP.fit}{The output of \code{SIMP()}.}

\item{Y}{Response matrix. Must have the same number of rows as \code{X}.}

\item{X1C}{Design matrix of the continuous part of the predictors of interest.
Must have the same number of rows as \code{Y}.}

\item{X1D}{Design matrix of the discrete part of the predictors of interest.
Must have the same number of rows as \code{Y}.}

\item{X2}{Design matrix of the nuisance predictors. Must have the same number of rows as \code{Y}.}

\item{dx}{Partial predictor envelope dimension. Must be an integer between 0 and \code{ncol(X1C)}.}

\item{dy}{Partial response envelope dimension. Must be an integer between 0 and \code{ncol(Y)}.}

\item{combined}{Logical. Indicate whether posterior samples (after burn-in) from all chains 
should be combined to calculate DIC. If there is only one chain, then there is no difference 
by indicating this parameter to be TRUE or FALSE.}

\item{Chain.no}{If combined = FALSE, we need to indicate which chain should be used for the 
calculation of DIC. Chain.no must be an integer between 1 and the number of chains.}
}
\description{
DIC calculation for the output of \code{SIMP()}
}
