% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_param.R
\name{elmwise_mean_in_list}
\alias{elmwise_mean_in_list}
\title{Compute the means of each entry in a matrix over a list}
\usage{
elmwise_mean_in_list(lst)
}
\arguments{
\item{lst}{A list of matrices of the same size.}
}
\description{
Compute the means of each entry in a matrix over a list
}
