% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_A_gamma.R
\name{find_A_from_gamma}
\alias{find_A_from_gamma}
\title{Function to compute A from from}
\usage{
find_A_from_gamma(gamma)
}
\arguments{
\item{gamma}{The basis matrix gamma of the envelope space.}
}
\description{
Function to compute A from from
}
