% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_A_gamma.R
\name{find_gammas_from_A}
\alias{find_gammas_from_A}
\title{Function to compute gamma from A}
\usage{
find_gammas_from_A(A)
}
\arguments{
\item{A}{The matrix A}
}
\description{
Function to compute gamma from A
}
