% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{find_reasonable_epsilon}
\alias{find_reasonable_epsilon}
\title{The function to find the stepsize}
\usage{
find_reasonable_epsilon(
  A,
  j,
  lpd_func,
  grad_lpd_func,
  M_diag,
  eps = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{A}{Parameter A or B.}

\item{j}{The index for which column of A or B to be updated.}

\item{lpd_func}{The log full conditional posterior density of A or B.}

\item{grad_lpd_func}{The function to calculate the gradients of 
the log full conditional posterior density of A or B.}

\item{M_diag}{Covariance matrix for our matrix parameter A or B.}

\item{eps}{The initial value for the stepsize}

\item{verbose}{Whether to show the progress messages.}

\item{...}{Other parameters that may be useful.}
}
\description{
The function to find the stepsize
}
