% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hyperparams_init.R
\name{get_hyperparams_init}
\alias{get_hyperparams_init}
\title{Hyper-parameters specification function}
\usage{
get_hyperparams_init(pc, pd, p2, r, dx, dy, ...)
}
\arguments{
\item{pc}{Dimension of X1C, the continuous part of the predictors of interest.}

\item{pd}{Dimension of X1D, the discrete part of the predictors of interest.}

\item{p2}{Dimension of X2, predictors of not main interest.}

\item{r}{Dimension of response Y.}

\item{dx}{Partial predictor envelope dimension of SIMP.}

\item{dy}{Partial response envelope dimension of SIMP.}

\item{...}{Input for the user-specified hyper-parameters.}
}
\description{
either generating a set of hyper-parameters for un-informative priors, or 
passing a set of user-specified hyper-parameters to \code{SIMP()} function.
}
\examples{
\dontrun{
library(SIMP)
pd = 8; pd = 2; p2 = 2; r = 8; dx = 1; dy = 1
hp <- get_hyperparams_init(pc, pd, p2, r, dx, dy)
}
}
