% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{grad_lpd_func}
\alias{grad_lpd_func}
\title{The function to calculate the gradients of the log full conditional posterior density of A or B.}
\usage{
grad_lpd_func(A, j, lpd_func, ...)
}
\arguments{
\item{A}{Parameter A or B}

\item{j}{The index for which column of A or B to be updated.}

\item{lpd_func}{The log full conditional posterior density of A or B.}

\item{...}{Other parameters that may be useful.}
}
\description{
The function to calculate the gradients of the log full conditional posterior density of A or B.
}
