% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMC.new.R
\name{HMC_new}
\alias{HMC_new}
\title{Hamiltonian monte-carlo}
\usage{
HMC_new(
  current_A,
  j,
  lpd_func,
  epsilon,
  LL,
  ColSigma.A.used = ColSigma.A.used,
  alpha,
  current_U,
  samp.size,
  ...
)
}
\arguments{
\item{current_A}{Current iterate of our matrix parameter A or B}

\item{j}{Index for which column to update}

\item{lpd_func}{Log full conditional posterior density}

\item{epsilon}{Step-size}

\item{LL}{Number of leapfrog steps in the trajectory}

\item{ColSigma.A.used}{Covariance matrix for our matrix parameter A or B}

\item{alpha}{Tempering parameter}

\item{current_U}{Energy function of the current iterate of our matrix parameter A or B}

\item{samp.size}{Sample size}

\item{...}{Other parameters needed for HMC updating.}
}
\description{
Hamiltonian monte-carlo
}
