% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{joint_log_density}
\alias{joint_log_density}
\title{The log joint full conditional density of the j-th column of A (or B) and its momentum vector.}
\usage{
joint_log_density(A, aa, lpd_func, M_diag, ...)
}
\arguments{
\item{A}{Parameter A or B.}

\item{aa}{Momentum vector of the j-th column of A (or B).}

\item{lpd_func}{The log full conditional posterior density of A or B.}

\item{M_diag}{Covariance matrix for our matrix parameter A or B.}

\item{...}{Other parameters that may be useful.}
}
\description{
The log joint full conditional density of the j-th column of A (or B) and its momentum vector.
}
