% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{leapfrog_step}
\alias{leapfrog_step}
\title{The function to implement the leapfrop step}
\usage{
leapfrog_step(A, aa, eps, grad_lpd_func, M_diag, j, lpd_func, ...)
}
\arguments{
\item{A}{Parameter A or B.}

\item{aa}{The momentum vector of the chosen column of A or B.}

\item{eps}{Stepsize.}

\item{grad_lpd_func}{The function to calculate the gradients of 
the log full conditional posterior density of A or B.}

\item{M_diag}{Covariance matrix for our matrix parameter A or B.}

\item{j}{The index for which column of A or B to be updated.}

\item{lpd_func}{The log full conditional posterior density of A or B.}

\item{...}{Other parameters that may be useful.}
}
\description{
The function to implement the leapfrop step
}
