% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpd.R
\name{lpd_A}
\alias{lpd_A}
\title{Function to compute the log full conditional density of A}
\usage{
lpd_A(
  A,
  X1C_ctr,
  X1C_ctr_shifted,
  Yctr,
  etaC,
  R,
  dy,
  pd,
  X1D_ctr_etaD.t,
  X2_ctr_beta2,
  Omega.inv,
  Omega0.inv,
  SigmaYX.inv,
  gamma_shifted,
  Lambda.half,
  A0,
  KA.half.inv,
  SigmaA.half.inv
)
}
\arguments{
\item{A}{Parameter A from previous iteration.}

\item{X1C_ctr}{Centered X1C.}

\item{X1C_ctr_shifted}{A inner variable.}

\item{Yctr}{Centered Y.}

\item{etaC}{Updated etaC in this iteration.}

\item{R}{R from previous iteration.}

\item{dy}{dy.}

\item{pd}{pd.}

\item{X1D_ctr_etaD.t}{A inner variable.}

\item{X2_ctr_beta2}{A inner variable.}

\item{Omega.inv}{Inverse of the updated Omega in this iteration.}

\item{Omega0.inv}{Inverse of the updated Omega0 in this iteration.}

\item{SigmaYX.inv}{Inverse of the updated SigmaYX in this iteration.}

\item{gamma_shifted}{A inner variable.}

\item{Lambda.half}{Inverse of Lambda.}

\item{A0}{A0.}

\item{KA.half.inv}{The (-1/2)-th power of KA.}

\item{SigmaA.half.inv}{The (-1/2)-th power of SigmaA.}
}
\description{
Function to compute the log full conditional density of A
}
