% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpd.R
\name{lpd_B}
\alias{lpd_B}
\title{Function to compute the log full conditional density of B}
\usage{
lpd_B(
  A,
  X1C_ctr,
  Yctr,
  L,
  dx,
  pd,
  etaC,
  X1D_ctr_etaD.t,
  X2_ctr_beta2,
  Phi.inv,
  Phi0.inv,
  beta2.shifted,
  M.half,
  B0,
  KB.half.inv,
  SigmaB.half.inv
)
}
\arguments{
\item{A}{Parameter B from previous iteration.}

\item{X1C_ctr}{Centered X1C.}

\item{Yctr}{Centered Y.}

\item{L}{Updated L.}

\item{dx}{dx}

\item{pd}{pd}

\item{etaC}{Updated etaC in this iteration.}

\item{X1D_ctr_etaD.t}{A inner variable.}

\item{X2_ctr_beta2}{A inner variable.}

\item{Phi.inv}{Inverse of the updated Phi in this iteration.}

\item{Phi0.inv}{Inverse of the updated Phi0 in this iteration.}

\item{beta2.shifted}{A inner variable.}

\item{M.half}{Square root of M.}

\item{B0}{B0}

\item{KB.half.inv}{The (-1/2)-th power of KB.}

\item{SigmaB.half.inv}{The (-1/2)-th power of SigmaB.}
}
\description{
Function to compute the log full conditional density of B
}
