% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{NUTS}
\alias{NUTS}
\title{The function to implement The No-U-Turn Sampler (NUTS)}
\usage{
NUTS(
  A,
  j,
  iter,
  lpd_func,
  grad_lpd_func,
  M_adapt,
  M_diag,
  delta = 0.6,
  max_treedepth = 10,
  eps = 1,
  eps_bar,
  H,
  mu,
  verbose = TRUE,
  Delta_max,
  ...
)
}
\arguments{
\item{A}{Parameter A or B.}

\item{j}{The index for which column of A or B to be updated.}

\item{iter}{The iteration index.}

\item{lpd_func}{The log full conditional posterior density of A or B.}

\item{grad_lpd_func}{The function to calculate the gradients of 
the log full conditional posterior density of A or B.}

\item{M_adapt}{The M_adapt parameter used in the Algorithm 6 of NUTS.}

\item{M_diag}{Covariance matrix for our matrix parameter A or B.}

\item{delta}{The delta parameter used in the Algorithm 6 of NUTS.}

\item{max_treedepth}{The maximal tree depth.}

\item{eps}{The initial value for the stepsize tuning for find_reasonable_epsilon().}

\item{eps_bar}{The eps_bar parameter used in the Algorithm 6 of NUTS.}

\item{H}{The H_m in the Algorithm 6 of NUTS.}

\item{mu}{The mu in the Algorithm 6 of NUTS.}

\item{verbose}{Whether to show the progress messages for find_reasonable_epsilon().}

\item{Delta_max}{The Delta_max parameter used in the Algorithm 6 of NUTS.}

\item{...}{Other parameters that may be useful.}
}
\description{
The function to implement The No-U-Turn Sampler (NUTS)
}
