% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rinvwish}
\alias{rinvwish}
\title{The function to generate a Inverse-Wishart sample}
\usage{
rinvwish(dim, Phi, nu)
}
\arguments{
\item{dim}{Dimension.}

\item{Phi}{Scale matrix parameter.}

\item{nu}{Degrees of freedom.}
}
\description{
The function to generate a Inverse-Wishart sample
}
