% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rMatrixNormal}
\alias{rMatrixNormal}
\title{The function to generate a Matrix Normal sample.}
\usage{
rMatrixNormal(M, V1, V2)
}
\arguments{
\item{M}{Location parameter.}

\item{V1}{Scale parameter for the row.}

\item{V2}{Scale parameter for the column.}
}
\description{
The function to generate a Matrix Normal sample.
}
