% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmh.colwise.new.R
\name{rmh_colwise_new}
\alias{rmh_colwise_new}
\title{Function to implement the Metropolis step.}
\usage{
rmh_colwise_new(
  A_start,
  lpd_func,
  method = "RW",
  tau = NULL,
  eps,
  eps_bar,
  H,
  mu,
  LL = 10,
  ColSigma.A.used,
  alpha,
  grad_lpd_func,
  M_adapt,
  M_diag,
  delta,
  iter,
  samp.size,
  ...
)
}
\arguments{
\item{A_start}{A or B in the current iteration.}

\item{lpd_func}{log full conditional density function of A or B.}

\item{method}{Method for the metropolis. We have three choices: RW (default), HMC, NUTS. We recommend to
use RW method always if no other reasons.}

\item{tau}{The standard deviation for the proposal distribution used in the RW-Metropolis.}

\item{eps}{Initial value for estimating the stepsize in NUTS-metropolis.}

\item{eps_bar}{eps_bar parameter for NUTS-metropolis.}

\item{H}{The H_m parameter for NUTS-metropolis.}

\item{mu}{The mu parameter for NUTS-metropolis.}

\item{LL}{Number of leapfrog steps in the trajectory}

\item{ColSigma.A.used}{Covariance matrix for our matrix parameter A or B}

\item{alpha}{Tempering parameter}

\item{grad_lpd_func}{The function to calculate the gradients of 
the log full conditional posterior density of A or B.}

\item{M_adapt}{The M_adapt parameter used in the Algorithm 6 of NUTS.}

\item{M_diag}{Covariance matrix for our matrix parameter A or B.}

\item{delta}{The delta parameter for NUTS-metropolis.}

\item{iter}{The iteration index.}

\item{samp.size}{Sample size.}

\item{...}{Other parameters that may be useful.}
}
\description{
Function to implement the Metropolis step.
}
