% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmvnorm}
\alias{rmvnorm}
\title{The function to generate vector Normal samples}
\usage{
rmvnorm(n = 1, mu, sigma, dim = NULL)
}
\arguments{
\item{n}{Sample size.}

\item{mu}{Mean.}

\item{sigma}{Covariance matrix.}

\item{dim}{Dimension.}
}
\description{
The function to generate vector Normal samples
}
