% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_acf.R
\name{sd_acf}
\alias{sd_acf}
\title{Function to calculate the standard deviations and the estimates for the auto-correlation function for 
posterior samples (PSD and PACF).}
\usage{
sd_acf(SIMP.fit, thinning_length = 1, acf.lagmax = 30)
}
\arguments{
\item{SIMP.fit}{Returned outputs from SIMP().}

\item{thinning_length}{The length for thinning before calculating PSD and PACF.}

\item{acf.lagmax}{The maximal lag length for calculating PACF.}
}
\description{
Function to calculate the standard deviations and the estimates for the auto-correlation function for 
posterior samples (PSD and PACF).
}
