% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{solve_chol}
\alias{solve_chol}
\title{An efficient function to calculate the inverse of a matrix}
\usage{
solve_chol(mat)
}
\arguments{
\item{mat}{The matrix that you want to calculate the inverse.}
}
\description{
An efficient function to calculate the inverse of a matrix
}
