% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqrtmat}
\alias{sqrtmat}
\title{The function to calculate the square root of a matrix}
\usage{
sqrtmat(mat)
}
\arguments{
\item{mat}{The matrix that you want to calculate the square root.}
}
\description{
The function to calculate the square root of a matrix
}
