% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqrtmatinv}
\alias{sqrtmatinv}
\title{The function to calculate the -(1/2)-th power of a matrix}
\usage{
sqrtmatinv(mat)
}
\arguments{
\item{mat}{The matrix that you want to calculate the -(1/2)-th power.}
}
\description{
The function to calculate the -(1/2)-th power of a matrix
}
